Add files via upload
This project will demonstrate how to use ImageView and use images in android studio. It will complete this objective by creating a dice using ImageView and after clicking on roll button, a random number will be generated and according to that random number generated between 0 to 6, image of that specific number will be shown.

That's' it.